wip page product-detail

+++ ecomm/toko/views.py
@@ -7,7 +7,6 @@ from django.utils import timezone
 from django.views import generic
 from paypal.standard.forms import PayPalPaymentsForm
 
-
 from .forms import CheckoutForm
 from .models import ProdukItem, OrderProdukItem, Order, AlamatPengiriman, Payment
 
@@ -20,17 +19,30 @@ class ProductDetailView(generic.DetailView):
     template_name = 'product_detail.html'
     queryset = ProdukItem.objects.all()
 
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        if self.request.user.is_authenticated:
+            product = self.get_object()
+            order_query = Order.objects.filter(user=self.request.user, ordered=False)
+            if order_query.exists():
+                order = order_query[0]
+                order_produk_item = order.produk_items.filter(produk_item=product).first()
+                if order_produk_item:
+                    product_quantity = order_produk_item.quantity
+                    context['product_quantity'] = product_quantity
+        return context
+
 class CheckoutView(LoginRequiredMixin, generic.FormView):
     def get(self, *args, **kwargs):
         form = CheckoutForm()
         try:
             order = Order.objects.get(user=self.request.user, ordered=False)
             if order.produk_items.count() == 0:
-                messages.warning(self.request, 'Belum ada belajaan yang Anda pesan, lanjutkan belanja')
+                messages.warning(self.request, 'You have not added anything to your cart')
                 return redirect('toko:home-produk-list')
         except ObjectDoesNotExist:
             order = {}
-            messages.warning(self.request, 'Belum ada belajaan yang Anda pesan, lanjutkan belanja')
+            messages.warning(self.request, 'You have not added anything to your cart')
             return redirect('toko:home-produk-list')
 
         context = {
@@ -66,10 +78,10 @@ class CheckoutView(LoginRequiredMixin, generic.FormView):
                 else:
                     return redirect('toko:payment', payment_method='stripe')
 
-            messages.warning(self.request, 'Gagal checkout')
+            messages.warning(self.request, 'Checkout failed')
             return redirect('toko:checkout')
         except ObjectDoesNotExist:
-            messages.error(self.request, 'Tidak ada pesanan yang aktif')
+            messages.error(self.request, 'No active orders')
             return redirect('toko:order-summary')
 
 class PaymentView(LoginRequiredMixin, generic.FormView):
@@ -113,7 +125,7 @@ class OrderSummaryView(LoginRequiredMixin, generic.TemplateView):
             template_name = 'order_summary.html'
             return render(self.request, template_name, context)
         except ObjectDoesNotExist:
-            messages.error(self.request, 'Tidak ada pesanan yang aktif')
+            messages.error(self.request, 'No active orders')
             return redirect('/')
 
 def add_to_cart(request, slug):
@@ -130,18 +142,18 @@ def add_to_cart(request, slug):
             if order.produk_items.filter(produk_item__slug=produk_item.slug).exists():
                 order_produk_item.quantity += 1
                 order_produk_item.save()
-                pesan = f"ProdukItem sudah diupdate menjadi: { order_produk_item.quantity }"
+                pesan = f"You have { order_produk_item.quantity } {order_produk_item.produk_item.nama_produk} in your cart"
                 messages.info(request, pesan)
                 return redirect('toko:produk-detail', slug = slug)
             else:
                 order.produk_items.add(order_produk_item)
-                messages.info(request, 'ProdukItem pilihanmu sudah ditambahkan')
+                messages.info(request, f'{order_produk_item.produk_item.nama_produk} has been added to your cart')
                 return redirect('toko:produk-detail', slug = slug)
         else:
             tanggal_order = timezone.now()
             order = Order.objects.create(user=request.user, tanggal_order=tanggal_order)
             order.produk_items.add(order_produk_item)
-            messages.info(request, 'ProdukItem pilihanmu sudah ditambahkan')
+            messages.info(request, f'{order_produk_item.produk_item.nama_produk} has been added to your cart')
             return redirect('toko:produk-detail', slug = slug)
     else:
         return redirect('/accounts/login')
@@ -156,25 +168,30 @@ def remove_from_cart(request, slug):
             order = order_query[0]
             if order.produk_items.filter(produk_item__slug=produk_item.slug).exists():
                 try: 
-                    order_produk_item = OrderProdukItem.objects.filter(
-                        produk_item=produk_item,
-                        user=request.user,
-                        ordered=False
-                    )[0]
-                    
-                    order.produk_items.remove(order_produk_item)
-                    order_produk_item.delete()
+                    # order_produk_item = OrderProdukItem.objects.filter(
+                    #     produk_item=produk_item,
+                    #     user=request.user,
+                    #     ordered=False
+                    # )[0]
+                    order_produk_item = order.produk_items.get(produk_item=produk_item, ordered=False)
+                    if order_produk_item.quantity > 1:
+                        order_produk_item.quantity -= 1
+                        order_produk_item.save()
+                        pesan = f"One {order_produk_item.produk_item.nama_produk} removed from cart"
+                    else:
+                        order.produk_items.remove(order_produk_item)
+                        order_produk_item.delete()
+                        pesan = f"{order_produk_item.produk_item.nama_produk} removed from cart"
 
-                    pesan = f"ProdukItem sudah dihapus"
                     messages.info(request, pesan)
                     return redirect('toko:produk-detail',slug = slug)
                 except ObjectDoesNotExist:
-                    print('Error: order ProdukItem sudah tidak ada')
+                    print(f'Error: {order_produk_item.produk_item.nama_produk} does not exist')
             else:
-                messages.info(request, 'ProdukItem tidak ada')
+                messages.info(request, f'This item does not exist in your cart')
                 return redirect('toko:produk-detail',slug = slug)
         else:
-            messages.info(request, 'ProdukItem tidak ada order yang aktif')
+            messages.info(request, f'No active orders of {order_produk_item.produk_item.nama_produk}')
             return redirect('toko:produk-detail',slug = slug)
     else:
         return redirect('/accounts/login')
@@ -200,17 +217,17 @@ def paypal_return(request):
             order.ordered = True
             order.save()
 
-            messages.info(request, 'Pembayaran sudah diterima, terima kasih')
+            messages.info(request, 'Payment recieved, thank you for your patronage')
             return redirect('toko:home-produk-list')
         except ObjectDoesNotExist:
-            messages.error(request, 'Periksa kembali pesananmu')
+            messages.error(request, 'Please check your orders again')
             return redirect('toko:order-summary')
     else:
         return redirect('/accounts/login')
 
 # @csrf_exempt
 def paypal_cancel(request):
-    messages.error(request, 'Pembayaran dibatalkan')
+    messages.error(request, 'Payment cancelled')
     return redirect('toko:order-summary')
 
 class products(generic.ListView):
